Update patch set 4
Patch Set 4:

> (7 comments)

Have made the changes. For the standardization, I want to keep the DataPipeline libraries in sync with protobuf based automatic build and top down design.
We can incorporate changes to this after the initial cut. As rthe code base stands right now it is an independent microservice

Patch-set: 4
Label: Code-Review=0